chore: move recording code to seperate file
